Remove `reverse` command.
Previously, `reverse` were used to generate a reverse patch for a
certain .ulp patch. Since now we have embedded the .ulp patch into
.so, there is no point in having a separate command to generate
a reverse patch, so we integrate it with packer.
